requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
pypdf
python-dotenv
sentence-transformers
langchain-chroma
langchain-huggingface
python-docx
fpdf
pymupdf
googlesearch-python
requests
duckduckgo-search
selenium
webdriver-manager
